feat: Remove fixed-size comments in CSS of index.html
Changes were made to the index.html CSS configuration. Though the body width and height are still fixed to 360px and 600px, respectively, and overflow is hidden, the comments in the CSS clearly indicating the setting of a fixed size for the body and hiding the overflow, have been removed.

This measure was taken to enhance readability of the code and eliminate unnecessary comments with self-evident information, streamlining codebase maintenance. This change doesn't alter the website's appearance or performance, but declutters the code and makes future updates logistically easier.

No associated issues.