feat: Update Visual Studio link in Home component
The reference URL in the Home component was updated from the old 'Theta Network' documentation link to the new Visual Studio Marketplace link for the 'Theta Edge X' extension. Now, users clicking on the Visual Studio icon in the Home page will be directly led to the relevant Visual Studio Marketplace page, enhancing the user experience and maintaining the relevance of referenced content.